Fixes for white label

+++ app/Http/Controllers/UserController.php
@@ -646,15 +646,19 @@ class UserController extends BaseController
      */
     public function detach(DetachCompanyUserRequest $request, User $user)
     {
+
         $company_user = CompanyUser::whereUserId($user->id)
                                     ->whereCompanyId(auth()->user()->companyId())->first();
                                     
         $token = $company_user->token->where('company_id', $company_user->company_id)->where('user_id', $company_user->user_id)->first();
 
-        $token->delete();
+        if($token)
+            $token->delete();
 
-        $company_user->delete();
+        if($company_user)
+            $company_user->delete();
 
         return response()->json(['message' => 'User detached from company'], 200);
+        
     }
 }
